+++ CHANGELOG.md
@@ -30,7 +30,8 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 -   Update telephonyEventHandler::getCall to parse more data from XMPP telephony event.
 -   Add Favorites service (fetchAllFavorites, createFavorite, deleteFavorite)
 -   Add Favorites events (rainbow_onfavoritecreated, rainbow_onfavoritedeleted)
-
+-   Add Favorites doc
+-   Add conversations::deleteMessage to delete a message by sending an empty string in a correctedMessage
 
 ## [1.57.0] - 2018-06-18
 -   RQRAINB-1550 : Add AdminService::getContactInfos method to retrieve informations about a user (need to be loggued in as a admin)